touch-ups per @rakuy0
