+++ vivisect/notes/switchcases.txt
